requirements.txt
# Flask Ticketing System - Production Requirements
# Core Flask packages
Flask==3.1.0
mysql-connector-python==9.4.0
Werkzeug==3.1.3

# Security and authentication
bcrypt==4.3.0

# Utilities
click==8.1.8
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0

# Email support (optional)
Flask-Mail==0.9.1

# Production server
gunicorn==23.0.0

# Environment variables
python-dotenv==1.1.1

# HTTP requests
requests==2.31.0

# Excel support
openpyxl==3.1.5
